+++ service/webApi/src/app.js
@@ -7,6 +7,15 @@ import { OpenAI } from 'openai'
 import pg from 'pg'
 import path from 'path'
 
+import * as action from './action.js'
+import * as core from './core.js'
+
+const asocial = {
+  core, action,
+}
+
+const a = asocial
+
 const PUBLIC_STATIC_DIR = 'view'
 const REGISTER_PROMPT = '/api/postMessage'
 const CHAT_HISTORY = '/api/getChatHistory'
@@ -15,6 +24,7 @@ const RENAME_CHAT = '/api/postRename'
 const DELETE_CHAT = '/api/postDelete'
 const TAG_LIST = '/api/getTagItemList'
 const REGISTER_CHAT_IN_TAG = '/api/registerChatInTag'
+const ADD_TAG = '/api/addTag'
 const DB_HOST = 'mermaid-chatgpt-editor-postgresql'
 const DB_PORT = 5432
 const DB_NAME = 'xl_db'
@@ -35,74 +45,50 @@ const _getDefaultRouter = () => {
   return expressRouter
 }
 
-const ADD_TAG = '/api/addTag'
 const _getFunctionRouter = () => {
   const expressRouter = express.Router()
 
-  // chatgpt
-  const registerPromptHandler = getHandlerRegisterPrompt({
-    handleRegisterPrompt: handleRegisterPrompt
+  const registerPromptHandler = a.action.getHandlerRegisterPrompt({
+    handleRegisterPrompt: a.core.handleRegisterPrompt
   })
   expressRouter.post(REGISTER_PROMPT, registerPromptHandler)
 
-  const chatHistoryHandler = getChatHistoryHandler({
-    handleChatHistory: handleChatHistory
+  const chatHistoryHandler = a.action.getHandlerChatHistory({
+    handleChatHistory: a.core.handleChatHistory
   })
   expressRouter.get(CHAT_HISTORY, chatHistoryHandler)
 
-  const chatListHandler = getChatListHandler({
-    handleChatList: handleChatList
+  const chatListHandler = a.action.getChatListHandler({
+    handleChatList: a.core.handleChatList
   })
   expressRouter.get(CHAT_LIST, chatListHandler)
 
-  const renameChatHandler = getHandlerRenameChat({
-    handleRenameChat: handleRenameChat
+  const renameChatHandler = a.action.getHandlerRenameChat({
+    handleRenameChat: a.core.handleRenameChat
   })
   expressRouter.post(RENAME_CHAT, renameChatHandler)
 
-  const deleteChatHandler = getHandlerDeleteChat({
-    handleDeleteChat: handleDeleteChat
+  const deleteChatHandler = a.action.getHandlerDeleteChat({
+    handleDeleteChat: a.core.handleDeleteChat
   })
   expressRouter.post(DELETE_CHAT, deleteChatHandler)
 
-  const tagItemListHandler = getHandlerTagItemList({
-    handleTagItemList: handleTagItemList
+  const tagItemListHandler = a.action.getHandlerTagItemList({
+    handleTagItemList: a.core.handleTagItemList
   })
   expressRouter.get(TAG_LIST, tagItemListHandler)
 
-  const addTagHandler = getHandlerAddTag({
-    handleAddTag: handleAddTag
+  const addTagHandler = a.action.getHandlerAddTag({
+    handleAddTag: a.core.handleAddTag
   })
   expressRouter.post(ADD_TAG, addTagHandler)
 
-  const registerChatInTagHandler = getHandlerRegisterChatInTag({
-    handleRegisterChatInTag: handleRegisterChatInTag
+  const registerChatInTagHandler = a.action.getHandlerRegisterChatInTag({
+    handleRegisterChatInTag: a.core.handleRegisterChatInTag
   })
   expressRouter.post(REGISTER_CHAT_IN_TAG, registerChatInTagHandler) 
 
   return expressRouter
-}
-
-const getHandlerRegisterChatInTag = ({ handleRegisterChatInTag }) => {
-  return async (req, res) => {
-    const { tagId, chatId } = req.body
-    console.log({ debug: true, request: 'registerChatInTag', tagId, chatId })
-
-    const result = await handleRegisterChatInTag({ tagId, chatId })
-
-    res.json({ result })
-  }
-}
-
-
-
-// チャットIDにタグを登録・更新する関数を実装
-const handleRegisterChatInTag = async ({ tagId, chatId }) => {
-  const paramList = [tagId, chatId]
-  const query = 'UPDATE chat_info.chat_list SET tag_id = $1 WHERE chat_id = $2'
-  const { result } = await execQuery({ query, paramList })
-  const { rowCount } = result
-  return rowCount === 1? 'ok': 'ng'
 }
 
 const _getErrorRouter = () => {
@@ -116,212 +102,6 @@ const _getErrorRouter = () => {
   return expressRouter
 }
 
-
-const handleAddTag = async ({ tagId, tagTitle }) => {
-  const paramList = [tagId, tagTitle]
-  const query = 'INSERT INTO chat_info.tag_list (tag_id, tag_title) VALUES ($1, $2)'
-  const { result } = await execQuery({ query, paramList })
-  const { rowCount } = result
-  return rowCount === 1? 'ok': 'ng'
-}
-
-
-// action
-const getHandlerAddTag = ({ handleAddTag }) => {
-  return async (req, res) => {
-    const { tagId, tagTitle } = req.body
-    console.log({ debug: true, request: 'addTag', tagId, tagTitle })
-
-    const result = await handleAddTag({ tagId, tagTitle })
-
-    res.json({ result })
-  }
-}
-
-const getHandlerRegisterPrompt = ({ handleRegisterPrompt }) => {
-  return async (req, res) => {
-    const { chatId, chatList } = req.body
-    console.log({ debug: true, request: 'ok!', chatId, chatList })
-
-    const handleResult = await handleRegisterPrompt({ chatId, chatList })
-
-    res.json({ result: handleResult })
-  }
-}
-
-const getChatHistoryHandler = ({ handleChatHistory }) => {
-  return async (req, res) => {
-    const { chatIdBefore } = req.query
-    console.log({ debug: true, request: 'ok!', chatIdBefore })
-
-    const handleResult = await handleChatHistory({ chatIdBefore })
-
-    res.json({ result: handleResult })
-  }
-}
-
-const getChatListHandler = ({ handleChatList }) => {
-  return async (req, res) => {
-    const { chatId } = req.query
-    console.log({ debug: true, request: 'ok!', chatId })
-
-    const handleResult = await handleChatList({ chatId })
-
-    res.json({ result: handleResult })
-  }
-}
-
-const getHandlerRenameChat = ({ handleRenameChat }) => {
-  return async (req, res) => {
-    const { chatId, oldChatTitle, newChatTitle } = req.body
-    console.log({ debug: true, request: 'ok!', chatId, oldChatTitle, newChatTitle })
-
-    const handleResult = await handleRenameChat({ chatId, oldChatTitle, newChatTitle })
-
-    res.json({ result: handleResult })
-  }
-}
-
-const getHandlerDeleteChat = ({ handleDeleteChat }) => {
-  return async (req, res) => {
-    const { chatId, chatTitle } = req.body
-    console.log({ debug: true, request: 'ok!', chatId, chatTitle })
-
-    const handleResult = await handleDeleteChat({ chatId, chatTitle })
-
-    res.json({ result: handleResult })
-  }
-}
-
-const getHandlerTagItemList = ({ handleTagItemList }) => {
-  return async (req, res) => {
-    console.log({ debug: true, request: 'GET Tag Item List' })
-
-    const handleResult = await handleTagItemList()
-
-    res.json({ result: handleResult })
-  }
-}
-
-const handleTagItemList = async () => {
-  const query = 'SELECT * FROM chat_info.chat_history h LEFT JOIN chat_info.tag_list p ON h.tag_id = p.tag_id where h.is_visible = true order BY p.tag_id DESC;'
-  const { result } = await execQuery({ query })
-  
-  const tagItemList = {}
-  result.rows.forEach((row) => {
-    const { tagId, tagTitle, chatId, chatTitle } = paramSnakeToCamel({ paramList: row })
-    if (!tagItemList[tagId]) {
-      tagItemList[tagId] = { tagTitle, historyList: [] }
-    }
-    tagItemList[tagId].historyList.push({ chatId, chatTitle })
-  })
-
-  return tagItemList
-}
-
-// core
-const handleRegisterPrompt = async ({ chatId, chatList }) => {
-  let currentChatId = null
-  console.log({ handleRegisterPrompt: true, chatId })
-  if (chatId) {
-    currentChatId = chatId
-  } else {
-    currentChatId = ulid()
-    await registerChatId({ chatId: currentChatId })
-  }
-
-  const prompt = chatList[chatList.length - 1].content
-  await registerChat({ chatId: currentChatId, role: 'user', content: prompt })
-
-  const stream = await mod.openaiClient.chat.completions.create({
-    model: 'gpt-4o',
-    // model: 'gpt-3.5-turbo',
-    messages: chatList,
-    stream: true,
-    max_tokens: 8192,
-  })
-  let responseMessage = ''
-  for await (const part of stream) {
-    // process.stdout.write(part.choices[0]?.delta?.content || '')
-    responseMessage += part.choices[0]?.delta?.content || ''
-  }
-
-  await registerChat({ chatId: currentChatId, role: 'assistant', content: responseMessage })
-
-  const handleResult = { chatId: currentChatId, message: responseMessage }
-  return handleResult
-}
-
-
-const registerChatId =  async ({ chatId }) => {
-  const date = new Date()
-  date.setHours(date.getHours() + 9)
-  const chatTitle = formatDate({ date })
-  const paramList = [chatId, chatTitle]
-  const query = 'insert into chat_info.chat_history (chat_id, chat_title) values ($1, $2)'
-  const { result } = await execQuery({ query, paramList })
-  const { rowCount } = result
-  return rowCount === 1? 'ok': 'ng'
-}
-
-const registerChat =  async ({ chatId, role, content }) => {
-  const paramList = [chatId, role, content]
-  const query = 'insert into chat_info.chat_list (chat_id, date_registered, role, content) values ($1, NOW(), $2, $3)'
-  const { result } = await execQuery({ query, paramList })
-  const { rowCount } = result
-  return rowCount === 1? 'ok': 'ng'
-}
-
-const handleRenameChat =  async ({ chatId, oldChatTitle, newChatTitle }) => {
-  const paramList = [newChatTitle, chatId, oldChatTitle]
-  const query = 'update chat_info.chat_history set chat_title = $1 where chat_id = $2 and chat_title = $3'
-  const { result } = await execQuery({ query, paramList })
-  const { affectedRows } = result
-  return affectedRows === 1? 'ok': 'ng'
-}
-
-const handleDeleteChat =  async ({ chatId, chatTitle }) => {
-  const paramList = [chatId, chatTitle]
-  const query = 'update chat_info.chat_history set is_visible = false where chat_id = $1 and chat_title = $2'
-  const { result } = await execQuery({ query, paramList })
-  const { affectedRows } = result
-  return affectedRows === 1? 'ok': 'ng'
-}
-
-
-const handleChatHistory = async ({ chatIdBefore }) => {
-  const paramList = []
-  let query = ''
-  if (chatIdBefore) {
-    query = 'select chat_id, chat_title from chat_info.chat_history where chat_id <= $1 and is_visible = true order by chat_id desc'
-    paramList.push(chatIdBefore)
-  } else {
-    query = 'select chat_id, chat_title from chat_info.chat_history where is_visible = true order by chat_id desc'
-  }
-  const { result } = await execQuery({ query, paramList })
-  const chatHistory = []
-  result.rows.forEach((row) => {
-    const { chatId, chatTitle } = paramSnakeToCamel({ paramList: row })
-    chatHistory.push({ chatId, chatTitle })
-  })
-
-  // return [{ chatId: 'abc', chatTitle: 'title abc', }, { chatId: 'xyz', chatTitle: 'title xyz', }]
-  return chatHistory
-}
-
-const handleChatList = async ({ chatId }) => {
-  const paramList = [chatId]
-  const query = 'select role, content from chat_info.chat_list where chat_id = $1 order by date_registered asc'
-  const { result } = await execQuery({ query, paramList })
-  const chatList = []
-  result.rows.forEach((row) => {
-    const { role, content } = row
-    chatList.push({ role, content })
-  })
-
-  // return [{"role":"user","content":"適当なmindmapをmermaidで出力して。投資に関するもの。"},{"role":"assistant","content":"もちろん、投資に関するマインドマップの例をMermaidで作成できます。以下に示します。\n\n```mermaid\nmindmap\n  root((投資))\n    資産クラス\n      株式\n        個別株\n        ETF\n        投資信託\n      債券\n        国債\n        社債\n      不動産\n        不動産投資信託(REIT)\n        不動産クラウドファンディング\n      商品\n        金\n        原油\n    投資スタイル\n      短期\n      中長期\n      デイトレード\n    リスク管理\n      分散投資\n      損切り\n      ヘッジ\n    分析手法\n      ファンダメンタルズ分析\n        株価収益率(PER)\n        株価純資産倍率(PBR)\n      テクニカル分析\n        移動平均線\n        RSI\n      マクロ経済分析\n        業種サイクル\n        経済指標\n    株式市場\n      国内市場\n      海外市場\n```\n\nこのマインドマップは、投資に関連する基本的な要素を示していますが、具体的な投資戦略や個別のニーズに応じて詳細を追加することもできます。"}]
-  return chatList
-}
 
 const startServer = ({ app, port }) => {
   app.listen(port, () => {
@@ -342,56 +122,6 @@ const createPgPool = ({ pg }) => {
   }
 
   return new pg.Pool(dbCredential)
-}
-
-const execQuery = async ({ query, paramList }) => {
-  if (paramList === undefined) {
-    paramList = []
-  }
-
-  return new Promise((resolve) => {
-    mod.pgPool
-      .query(query, paramList)
-      .then((result) => {
-        console.log({ query, paramList, rows: result.rows })
-        return resolve({ err: null, result })
-      })
-      .catch((err) => {
-        console.log('Error executing query', { err })
-        return resolve({ err, result: null })
-      })
-  })
-}
-
-const formatDate = ({ format, date }) => {
-  if (format === undefined) {
-    format = 'YYYY-MM-DD hh:mm:ss'
-  }
-  if (date === undefined) {
-    date = new Date()
-  }
-
-  return format.replace(/YYYY/g, date.getFullYear())
-    .replace(/MM/g, (`0${date.getMonth() + 1}`).slice(-2))
-    .replace(/DD/g, (`0${date.getDate()}`).slice(-2))
-    .replace(/hh/g, (`0${date.getHours()}`).slice(-2))
-    .replace(/mm/g, (`0${date.getMinutes()}`).slice(-2))
-    .replace(/ss/g, (`0${date.getSeconds()}`).slice(-2))
-}
-
-const paramSnakeToCamel = ({ paramList }) => {
-  if (paramList === undefined) {
-    paramList = {}
-  }
-
-  const newParamList = {}
-  Object.entries(paramList).forEach(([key, value]) => {
-    const newKey = key.replace(/([_][a-z])/g, (group) => {
-      return group.toUpperCase().replace('_', '')
-    })
-    newParamList[newKey] = value
-  })
-  return newParamList
 }
 
 const init = async () => {
